backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
pymongo==4.9.1
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
whitenoise==6.7.0
gunicorn==23.0.0
uvicorn==0.30.1
resend==2.4.0
requests==2.32.3
Pillow==10.4.0
djongo==1.3.6
mongodb-db-engine
